clay(P19): make Clay the default lobby; delete legacy lobby files
Rebases LobbyClayScreen onto Screen directly (was LobbyScreen) and
inlines the entire lobby pump (DISCONNECTED bounce, joininggame
finalisation, progress-modal spinner, CONNECTED handoff,
mapDownloader.ProcessMapDownload, disconnect modal) into its Tick.
Game::GoToState(LOBBY) pushes LobbyClayScreen unconditionally — the
SILENCER_LOBBY_CLAY getenv check is gone.

Deletes the legacy LobbyScreen + its 6 panel classes (~2440 LOC):
- clients/silencer/src/ui/screens/lobby/lobby_screen.{h,cpp}
- clients/silencer/src/ui/screens/lobby/panels/{character,chat,
  game_select,game_create,game_join,game_tech}_panel.{h,cpp}

Updates controldispatch.cpp's lobby_show_panel op to cast to
LobbyClayScreen *, world.h's friend list to drop the deleted
classes, CMakeLists.txt to drop the legacy include dirs, and the
40_lobby_clay_basic e2e to drop the now-no-op env var.

Pass check: cmake build green; tests/cli-agent/run.sh all 5 green;
all 7 tests/lobby-clay/*/run.sh under 0.65% (well below 5%).
